+++ src/ThinkOnErp.Application/Features/Users/Queries/GetUserById/GetUserByIdQuery.cs
@@ -5,5 +5,5 @@ namespace ThinkOnErp.Application.Features.Users.Queries.GetUserById;
 
 public class GetUserByIdQuery : IRequest<UserDto?>
 {
-    public decimal RowId { get; set; }
+    public Int64 RowId { get; set; }
 }
