+++ src/ThinkOnErp.Domain/Entities/SysCurrency.cs
@@ -10,7 +10,7 @@ public class SysCurrency
     /// <summary>
     /// Primary key - generated from SEQ_SYS_CURRENCY sequence
     /// </summary>
-    public decimal RowId { get; set; }
+    public Int64 RowId { get; set; }
 
     /// <summary>
     /// Arabic description of the currency
